+++ examples/resources/iosxr_flow_monitor_map/resource.tf
@@ -1,8 +1,8 @@
 resource "iosxr_flow_monitor_map" "example" {
-  monitor_map_name = "monitor_map1"
+  name = "monitor_map1"
   exporters = [
     {
-      exporter_name = "exporter1"
+      name = "exporter1"
     }
   ]
   option_outphysint                          = true
@@ -31,7 +31,7 @@ resource "iosxr_flow_monitor_map" "example" {
   record_mpls_ipv4_fields                    = true
   record_mpls_ipv6_fields                    = true
   record_mpls_ipv4_ipv6_fields               = true
-  record_mpls_labels                         = true
+  record_mpls_labels                         = 2
   record_map_t                               = true
   record_sflow                               = true
   record_datalink_record                     = true
@@ -51,6 +51,6 @@ resource "iosxr_flow_monitor_map" "example" {
   sflow_options_extended_ipv6_tunnel_egress  = true
   sflow_options_if_counters_polling_interval = 5
   sflow_options_sample_header_size           = 128
-  sflow_options_input_ifindex                = "index1"
-  sflow_options_output_ifindex               = "index2"
+  sflow_options_input_ifindex                = "physical"
+  sflow_options_output_ifindex               = "physical"
 }