+++ internal/provider/resource_iosxr_flow_monitor_map_test.go
@@ -27,9 +27,12 @@ import (
 )
 
 func TestAccIosxrFlowMonitorMap(t *testing.T) {
+	if os.Getenv("FLOW") == "" {
+		t.Skip("skipping test, set environment variable FLOW")
+	}
 	var checks []resource.TestCheckFunc
-	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "monitor_map_name", "monitor_map1"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "exporters.0.exporter_name", "exporter1"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "name", "monitor_map1"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "exporters.0.name", "exporter1"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "option_outphysint", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "option_filtered", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "option_bgpattr", "true"))
@@ -56,7 +59,7 @@ func TestAccIosxrFlowMonitorMap(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_mpls_ipv4_fields", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_mpls_ipv6_fields", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_mpls_ipv4_ipv6_fields", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_mpls_labels", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_mpls_labels", "2"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_map_t", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_sflow", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "record_datalink_record", "true"))
@@ -76,8 +79,8 @@ func TestAccIosxrFlowMonitorMap(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "sflow_options_extended_ipv6_tunnel_egress", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "sflow_options_if_counters_polling_interval", "5"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "sflow_options_sample_header_size", "128"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "sflow_options_input_ifindex", "index1"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "sflow_options_output_ifindex", "index2"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "sflow_options_input_ifindex", "physical"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxr_flow_monitor_map.test", "sflow_options_output_ifindex", "physical"))
 	var steps []resource.TestStep
 	if os.Getenv("SKIP_MINIMUM_TEST") == "" {
 		steps = append(steps, resource.TestStep{
@@ -102,16 +105,16 @@ func TestAccIosxrFlowMonitorMap(t *testing.T) {
 
 func testAccIosxrFlowMonitorMapConfig_minimum() string {
 	config := `resource "iosxr_flow_monitor_map" "test" {` + "\n"
-	config += `	monitor_map_name = "monitor_map1"` + "\n"
+	config += `	name = "monitor_map1"` + "\n"
 	config += `}` + "\n"
 	return config
 }
 
 func testAccIosxrFlowMonitorMapConfig_all() string {
 	config := `resource "iosxr_flow_monitor_map" "test" {` + "\n"
-	config += `	monitor_map_name = "monitor_map1"` + "\n"
+	config += `	name = "monitor_map1"` + "\n"
 	config += `	exporters = [{` + "\n"
-	config += `		exporter_name = "exporter1"` + "\n"
+	config += `		name = "exporter1"` + "\n"
 	config += `	}]` + "\n"
 	config += `	option_outphysint = true` + "\n"
 	config += `	option_filtered = true` + "\n"
@@ -139,7 +142,7 @@ func testAccIosxrFlowMonitorMapConfig_all() string {
 	config += `	record_mpls_ipv4_fields = true` + "\n"
 	config += `	record_mpls_ipv6_fields = true` + "\n"
 	config += `	record_mpls_ipv4_ipv6_fields = true` + "\n"
-	config += `	record_mpls_labels = true` + "\n"
+	config += `	record_mpls_labels = 2` + "\n"
 	config += `	record_map_t = true` + "\n"
 	config += `	record_sflow = true` + "\n"
 	config += `	record_datalink_record = true` + "\n"
@@ -159,8 +162,8 @@ func testAccIosxrFlowMonitorMapConfig_all() string {
 	config += `	sflow_options_extended_ipv6_tunnel_egress = true` + "\n"
 	config += `	sflow_options_if_counters_polling_interval = 5` + "\n"
 	config += `	sflow_options_sample_header_size = 128` + "\n"
-	config += `	sflow_options_input_ifindex = "index1"` + "\n"
-	config += `	sflow_options_output_ifindex = "index2"` + "\n"
+	config += `	sflow_options_input_ifindex = "physical"` + "\n"
+	config += `	sflow_options_output_ifindex = "physical"` + "\n"
 	config += `}` + "\n"
 	return config
 }