+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `timers_bgp_minimum_acceptable_holdtime` attribute to `iosxr_router_bgp` resource and data source
 - Add `iosxr_flow_sampler_map` resource and data source
+- Add `iosxr_flow_monitor_map` resource and data source
 
 ## 0.2.2